.long()

+++ manydepth/trainer_m2_normals.py
@@ -768,7 +768,7 @@ class Trainer_Monodepth2:
         bottom_left_depth = bottom_left_flat.permute(0, 2, 1).to(device=K_inv.device) * D.view(batch_size, 1, -1)
         #print(top_left_flat.shape)
         #top_left_depth = torch.gather(D, 2, top_left_flat.unsqueeze(1).long())
-        top_left_depth = D[:, :, top_left_flat[:, :, 1], top_left_flat[:, :, 0]]
+        top_left_depth = D[:, :, top_left_flat[:, :, 1].long(), top_left_flat[:, :, 0].long()]
         print(top_left_depth.shape)
         #print(ones.shape)
